refactor USerTExt
